+++ src/main/java/com/jarias14/tekstratego/service/thinker/model/Study.java
@@ -1,5 +1,6 @@
 package com.jarias14.tekstratego.service.thinker.model;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.Map;
 import java.util.SortedMap;
@@ -21,7 +22,7 @@ public interface Study {
     
     public void setId(String id);
     public String getId();
-    public boolean execute(Map<String,SortedMap<Calendar, Double>> indicatorValues);
+    public boolean execute(Map<String,SortedMap<Calendar, BigDecimal>> indicatorValues);
     public StudyResource toResource();
     public String getParentId();
     public void setParentId(String id);
